Revert "Simplify entrypoint code (#352)"
This reverts commit 56dec8990375cb2af7a4f1b7d9046f93387442fe.
