+++ layer/context.py
@@ -1,7 +1,7 @@
 from types import TracebackType
 from typing import Optional
 
-from layer.contracts.asset import AssetType
+from layer.contracts.assets import AssetType
 from layer.contracts.datasets import DatasetBuild
 from layer.tracker.ui_progress_tracker import RunProgressTracker
 from layer.training.base_train import BaseTrain
