+++ layer/contracts/assets.py
@@ -1,0 +1,5 @@
+"""
+This module is here to facilitate the transition to the new module name without breaking compatibility between the SDK and the backend
+"""
+# pylint: disable=unused-import
+from .asset import AssetPath, AssetType, BaseAsset  # noqa
