+++ test/unit/decorators/test_dataset_decorator.py
@@ -8,7 +8,7 @@ from layerapi.api.ids_pb2 import DatasetBuildId
 from layerapi.api.service.datacatalog.data_catalog_api_pb2 import InitiateBuildResponse
 
 from layer.clients.data_catalog import DataCatalogClient
-from layer.contracts.asset import AssetType
+from layer.contracts.assets import AssetType
 from layer.contracts.datasets import Dataset
 from layer.contracts.fabrics import Fabric
 from layer.contracts.models import Model
